compilation of pcl_featurs under win64
minor warning

git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@3244 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ cuda/containers/include/pcl/gpu/containers/device_array.hpp
@@ -255,7 +255,7 @@ namespace pcl
             int rows() const;
 
             /** \brief Returns step in elements. */
-            int elem_step() const;
+            size_t elem_step() const;
         };        
     }
