+++ cuda/utils/include/pcl/gpu/utils/device/funcattrib.hpp
@@ -52,9 +52,9 @@ namespace pcl
 
             printf("=== Function stats ===\n");
             printf("Name: \n");
-            printf("sharedSizeBytes    = %d\n", attrs.sharedSizeBytes);
-            printf("constSizeBytes     = %d\n", attrs.constSizeBytes);
-            printf("localSizeBytes     = %d\n", attrs.localSizeBytes);
+            printf("sharedSizeBytes    = %d\n", (int)attrs.sharedSizeBytes);
+            printf("constSizeBytes     = %d\n", (int)attrs.constSizeBytes);
+            printf("localSizeBytes     = %d\n", (int)attrs.localSizeBytes);
             printf("maxThreadsPerBlock = %d\n", attrs.maxThreadsPerBlock);
             printf("numRegs            = %d\n", attrs.numRegs);
             printf("ptxVersion         = %d\n", attrs.ptxVersion);
